3.3.0 b565
* changed custom drops and rules that use role-chance logic to use a different formula and float datatype so that small chance numbers work as expected
* fixed bug that prevented nbt-data being parsed correctly when used in a custom rule
* when using nbt-data and have the 'NBT_APPLY_SUCCESS' debug enabled, it will show what fields were changed/added/removed

+++ src/main/java/me/lokka30/levelledmobs/customdrops/CustomDropBase.java
@@ -37,7 +37,7 @@ public class CustomDropBase implements Cloneable {
     public int maxDropGroup;
     public int minPlayerLevel;
     public int maxPlayerLevel;
-    public double chance;
+    public float chance;
     public boolean playerCausedOnly;
     public boolean noSpawner;
     public String groupId;
